feat : running result 화면 backstack 관리
