+++ RunWithMe/presentation/src/main/java/com/side/runwithme/view/running_result/RunningResultViewModel.kt
@@ -52,6 +52,6 @@ class RunningResultViewModel @Inject constructor(
     }
 
     fun putCoordinates(coordinates: List<Coordinate>){
-
+        _coordinates.value = coordinates
     }
 }
